dfu-util: Add option to force the existence and correct DFU suffix
While we had the DFU suffix parsing and checking in place for some
time now we did not enforce it. With this patch we still don't do it
by default, but we allow to enforce the suffix with -f or --force-suffix

I got a request for this from a real world customer having problems with
people flashing corrupted files to the devices.